NonMovingScav: Account for new allocations
Previously if we evacuated any objects to the segment that we were
scavenging we would exit the scavenge loop before having scavenged them.
This isn't a correctness issue since we would try again to scavenge
later but it is slightly more efficient to handle newly-allocated
objects ASAP after we allocate them.